requirements.txt
django==5.2
djangorestframework==3.16.0
python-jose==3.4.0
psycopg2-binary==2.9.10
redis==5.2.1
openai==1.73.0
gunicorn==21.2.0
whitenoise==6.6.0
django-cors-headers==4.3.1
PyJWT==2.8.0